Change Command to an abstract class and add its subclasses
