+++ src/main/java/commands/ListCommand.java
@@ -1,0 +1,15 @@
+package commands;
+
+import core.TaskList;
+
+public class ListCommand extends Command {
+    @Override
+    public void execute(TaskList taskList) {
+        taskList.listTasks();
+    }
+
+    @Override
+    public boolean shouldExit() {
+        return false;
+    }
+}
